# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MDGA Enterprise B2B SaaS",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/ingest": {
      "post": {
        "tags": [
          "Ingest v1"
        ],
        "summary": "Ingest",
        "operationId": "ingest_api_v1_ingest_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_ingest_api_v1_ingest_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/chat": {
      "post": {
        "tags": [
          "Chat v1"
        ],
        "summary": "Chat With Copilot",
        "operationId": "chat_with_copilot_api_v1_chat_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatPayload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Body_ingest_api_v1_ingest_post": {
        "properties": {
          "raw_text": {
            "type": "string",
            "title": "Raw Text"
          },
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "location": {
            "type": "string",
            "title": "Location"
          },
          "is_guest": {
            "type": "string",
            "title": "Is Guest",
            "default": "false"
          },
          "industry": {
            "type": "string",
            "title": "Industry",
            "default": "\uacf5\uacf5"
          }
        },
        "type": "object",
        "required": [
          "location"
        ],
        "title": "Body_ingest_api_v1_ingest_post"
      },
...
      "ChatPayload": {
        "properties": {
          "path": {
            "type": "string",
            "title": "Path"
          },
          "industry": {
            "type": "string",
            "title": "Industry"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "selected_hashes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Selected Hashes",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "path",
          "industry",
          "message"
        ],
        "title": "ChatPayload"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}